+++ src/TestSuite/TestTrait.php
@@ -16,6 +16,7 @@ namespace Tools\TestSuite;
 use BadMethodCallException;
 use Exception;
 use PHPUnit\Framework\Constraint\IsType;
+use PHPUnit\Framework\Error\Error;
 
 /**
  * A trait that provides some assertion methods
@@ -128,6 +129,8 @@ trait TestTrait
 
         try {
             call_user_func($function);
+        } catch (Error $e) {
+            throw $e;
         } catch (Exception $e) {
             parent::assertInstanceof(
                 $expectedException,